Get project data

Import project-level data from googlesheets.

pandan_dat(project, gs_url = Sys.getenv("PANDAN_TRACKER"))

Arguments

project
Project tracking sheet. First column for components, second to some number of columns for project level progress, and any column name containing read_edit will be classified as reading over and editing, not project progress.

gs_url
Url for googlesheet. Defaults to system environment variable PANDAN. usethis::edit_r_environ() to set this.

Value

List of object of project metadata useful for visualisation, e.g., current levels and current sections.

Examples

# runs because PANDAN url environment variable set and
# pandan googlesheet at url
pandan_dat("waystation")

#> → Using an auto-discovered, cached token

#>   To suppress this message, modify your code or options to clearly consent to
#>   the use of a cached token

#>   See gargle's "Non-interactive auth" vignette for more details:

#>   <https://gargle.r-lib.org/articles/non-interactive-auth.html>

#> → The googlesheets4 package is using a cached token for charlestigray@gmail.com

#> Auto-refreshing stale OAuth token.

#> Reading from "pandan trackers"

#> Range "'waystation'"

#> New names:
#> * read_edit -> read_edit...6
#> * read_edit -> read_edit...7
#> * read_edit -> read_edit...8

#> $components
#> # A tibble: 10 x 2
#>    components completed
#>    <chr>          <int>
#>  1 1                  4
#>  2 1.1                3
#>  3 1.2                3
#>  4 2                  4
#>  5 2.1                3
#>  6 2.2                3
#>  7 2.2.1              2
#>  8 2.2.2              2
#>  9 2.2.3              2
#> 10 3                  3
#> 
#> $editing
#> # A tibble: 10 x 2
#>    components completed
#>    <chr>          <int>
#>  1 1                  1
#>  2 1.1                1
#>  3 1.2                1
#>  4 2                  1
#>  5 2.1                1
#>  6 2.2                0
#>  7 2.2.1              0
#>  8 2.2.2              0
#>  9 2.2.3              0
#> 10 3                  0
#> 
#> $sections
#>  [1] "1"     "1.1"   "1.2"   "2"     "2.1"   "2.2"   "2.2.1" "2.2.2" "2.2.3"
#> [10] "3"    
#> 
#> $levels
#> [1] "heading"   "sentence"  "paragraph" "section"  
#> 
#> $edit_n
#> [1] 3
#> 
#> attr(,"class")
#> [1] "pandan_project"